UR: Se mejoró la logica para el ciclo de vida de los jwt en login y logout V2

+++ src/controllers/logout.ts
@@ -1,10 +1,6 @@
 import { Request, Response } from "express";
 import asyncHandler from "express-async-handler";
-import jwt from "jsonwebtoken";
 import { db } from "../lib/db";
-
-// Asegúrate de que JWT_SECRET esté definido en tu archivo de configuración de entorno
-const JWT_SECRET = process.env.JWT_SECRET || 'Un1R0Om202A*@*';
 
 export const logout = asyncHandler(async (req: Request, res: Response): Promise<void> => {
   const { token } = req.body;
@@ -14,20 +10,29 @@ export const logout = asyncHandler(async (req: Request, res: Response): Promise<
     return;
   }
 
-  // Decodificar el token para obtener el userId
-  try {
-    const decodedToken: any = jwt.verify(token, JWT_SECRET);
+  // Buscar el token en la base de datos y obtener el userId
+  const jwtTokenRecord = await db.jwtToken.findUnique({
+    where: { token: token },
+  });
 
-    const userId = decodedToken.userId;
-
-    // Eliminar todos los tokens asociados con este usuario
-    await db.jwtToken.deleteMany({
-      where: { userId: userId },
-    });
-
-    // Notificar al usuario que todos los JWT se han eliminado y la sesión ha sido cerrada
-    res.json({ success: true, message: "Sesión cerrada" });
-  } catch (error) {
-    res.status(400).json({ error: "Token inválido o expirado" });
+  if (!jwtTokenRecord) {
+    res.status(404).json({ error: "Token no encontrado" });
+    return;
   }
+
+  const userId = jwtTokenRecord.userId;
+
+  // Eliminar todos los tokens asociados con este userId
+  const deletedTokens = await db.jwtToken.deleteMany({
+    where: { userId: userId },
+  });
+
+  // Verifica si se eliminaron tokens
+  if (deletedTokens.count === 0) {
+    res.status(404).json({ error: "No se encontraron tokens para este usuario" });
+    return;
+  }
+
+  // Notificar al usuario que todos los JWT se han eliminado y la sesión ha sido cerrada
+  res.json({ success: true, message: "Sesión cerrada y todos los tokens eliminados" });
 });
